+++ messaginginapp/src/main/java/io/customer/messaginginapp/ModuleMessagingInApp.kt
@@ -2,9 +2,11 @@ package io.customer.messaginginapp
 
 import io.customer.messaginginapp.di.gistCustomAttributes
 import io.customer.messaginginapp.di.gistProvider
+import io.customer.messaginginapp.di.notificationInbox
 import io.customer.messaginginapp.gist.data.model.Message
 import io.customer.messaginginapp.gist.presentation.GistListener
 import io.customer.messaginginapp.gist.presentation.GistProvider
+import io.customer.messaginginapp.inbox.NotificationInbox
 import io.customer.messaginginapp.type.InAppMessage
 import io.customer.sdk.communication.Event
 import io.customer.sdk.communication.subscribe
@@ -22,6 +24,15 @@ class ModuleMessagingInApp(
     private val gistProvider: GistProvider
         get() = SDKComponent.gistProvider
     private val logger = SDKComponent.logger
+
+    /**
+     * Access the inbox messages instance for managing user inbox messages.
+     *
+     * @return [NotificationInbox] instance for inbox operations
+     */
+    fun inbox(): NotificationInbox {
+        return SDKComponent.notificationInbox
+    }
 
     fun dismissMessage() {
         gistProvider.dismissMessage()