+++ messaginginapp/src/main/java/io/customer/messaginginapp/gist/data/model/response/InboxMessageExtensions.kt
@@ -1,0 +1,6 @@
+package io.customer.messaginginapp.gist.data.model.response
+
+import io.customer.messaginginapp.gist.data.model.InboxMessage
+
+// Formats inbox message for logging.
+internal fun InboxMessage.toLogString(): String = "$queueId (deliveryId: $deliveryId)"