+++ messaginginapp/src/main/java/io/customer/messaginginapp/gist/data/sse/InAppSseLogger.kt
@@ -83,8 +83,8 @@ internal class InAppSseLogger(private val logger: Logger) {
         logger.debug(tag = TAG, message = "Received heartbeat")
     }
 
-    fun logReceivedMessages(count: Int) {
-        logger.debug(tag = TAG, message = "Received $count messages")
+    fun logReceivedMessages(count: Int, type: String?) {
+        logger.debug(tag = TAG, message = "Received $count $type messages")
     }
 
     fun logReceivedEmptyMessagesEvent() {
@@ -217,6 +217,10 @@ internal class InAppSseLogger(private val logger: Logger) {
 
     fun logHeartbeatTimeoutParsingError(errorMessage: String?, data: String) {
         logger.debug(tag = TAG, message = "Error parsing heartbeat timeout: $errorMessage, data: $data")
+    }
+
+    fun logFilteredInvalidInboxMessages(count: Int) {
+        logger.debug(tag = TAG, message = "Filtered out $count invalid inbox message(s) from SSE")
     }
 
     // =====================