+++ samples/java_layout/src/main/java/io/customer/android/sample/java_layout/ui/dashboard/DashboardActivity.java
@@ -32,6 +32,7 @@ import io.customer.android.sample.java_layout.R;
 import io.customer.android.sample.java_layout.databinding.ActivityDashboardBinding;
 import io.customer.android.sample.java_layout.sdk.CustomerIORepository;
 import io.customer.android.sample.java_layout.ui.core.BaseActivity;
+import io.customer.android.sample.java_layout.ui.inbox.InboxMessagesActivity;
 import io.customer.android.sample.java_layout.ui.inline.InlineExamplesActivity;
 import io.customer.android.sample.java_layout.ui.location.LocationTestActivity;
 import io.customer.android.sample.java_layout.ui.login.LoginActivity;
@@ -156,6 +157,9 @@ public class DashboardActivity extends BaseActivity<ActivityDashboardBinding> {
         binding.inlineExamplesButton.setOnClickListener(view -> {
             startInlineExamplesActivity();
         });
+        binding.inboxMessagesButton.setOnClickListener(view -> {
+            startInboxMessagesActivity();
+        });
         binding.logoutButton.setOnClickListener(view -> {
             authViewModel.clearLoggedInUser();
         });
@@ -206,6 +210,11 @@ public class DashboardActivity extends BaseActivity<ActivityDashboardBinding> {
 
     private void startInlineExamplesActivity() {
         Intent intent = new Intent(DashboardActivity.this, InlineExamplesActivity.class);
+        startActivity(intent);
+    }
+
+    private void startInboxMessagesActivity() {
+        Intent intent = new Intent(DashboardActivity.this, InboxMessagesActivity.class);
         startActivity(intent);
     }